feat: write monitoring data periodically; improve logging

+++ mama-module/monitoring/start_monitors.sh
@@ -1,10 +1,17 @@
 #!/usr/bin/env bash
 
+INEQUALITY_INPUTS_PATH="$(realpath $1)"
+if [ -z "$INEQUALITY_INPUTS_PATH" ]; then
+  echo "Usage: $0 <inequality_inputs_path>"
+  exit 1
+fi
+echo "INEQUALITY_INPUTS_PATH = $INEQUALITY_INPUTS_PATH"
+
 rm -f monitor*.log
 rm -f totalRates*.log
 for i in {0..4}
 do
-  if [ $i -ne 2 ] && [ $i -ne 3 ]; then
-		java -ea --add-opens java.base/java.lang=ALL-UNNAMED -jar target/monitoring-0.1.jar -node $i -monitoringinputs "/Users/krispian/Uni/bachelorarbeit/test_flink_inputs_dev/generate_flink_inputs/plans/inequality_inputs.json" > monitor$i.log 2>&1 &
-  fi
+  # if [ $i -ne 1 ] ; then
+		java -ea --add-opens java.base/java.lang=ALL-UNNAMED -jar target/monitoring-0.1.jar -node $i -monitoringinputs $INEQUALITY_INPUTS_PATH > monitor$i.log 2>&1 &
+  # fi
 done
